updated makefile and moved helper functions to separate file, need to fix removeFile function
