+++ src/tests/test_1.cpp
@@ -1,6 +1,9 @@
 #include "../rt/rt.hpp"
+#include "../rt/helper.hpp"
 
 int main() {
+    removeFile("table1.tbl");
+
     RelationalTable a("table1.tbl", 69);
     std::cout << a.readNumEntries() << std::endl;
     std::cout << a.readNumColumns() << std::endl;
